+++ internal/entity/doc.go
@@ -1,0 +1,5 @@
+// Package entity provides internal implementations of the entities defined in
+// the configkit package.
+//
+// deprecated: Packages should define their own implementations instead.
+package entity