+++ message/type_test.go
@@ -100,6 +100,16 @@ var _ = Describe("type Type", func() {
 		})
 	})
 
+	Describe("func TypeFor()", func() {
+		It("returns the type", func() {
+			tb := TypeOf(fixtures.MessageA1)
+			ta := TypeFor[fixtures.MessageA]()
+
+			Expect(ta).To(Equal(tb))
+			Expect(ta == tb).To(BeTrue()) // we're testing == here specifically, hence not using To(Equal())
+		})
+	})
+
 	Describe("func TypeFromReflect()", func() {
 		It("returns values that compare as equal for messages of the same type", func() {
 			tb := TypeFromReflect(reflect.TypeOf(fixtures.MessageA1))