+++ CHANGELOG.md
@@ -10,6 +10,16 @@ The format is based on [Keep a Changelog], and this project adheres to
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
+## [Unreleased]
+
+### Added
+
+- Added `message.TypeFor()`.
+
+### Changed
+
+- Bumped minimum Go version to 1.22.
+
 ## [0.13.3] - 2024-08-15
 
 ### Added